+++ docs/source/changelog.rst
@@ -1,6 +1,12 @@
 ChangeLog
 ---------
 
+**v2.5.1** (2022-04-03)
+
+- Class TraversalShortestPaths: Error in documentation corrected,
+  option keep_distances added.
+
+
 **v2.5.0** (2022-03-28)
 
 - First public version marked as stable (PyPI and GitHub)